fix: исправлены ошибки, добавлена кнопка для просмотра пароля
1) Внесены исправления в стили и названия классов
2)Изменено расположение файлов
3)Добавлена функция просмотра вводимого пароля
