+++ src/Widgets/AuthorizationWidget/index.tsx
@@ -1,0 +1,2 @@
+import { AuthorizationPage } from './AuthorizationWidget';
+export default AuthorizationPage;
