--- src/pages/AuthorizationPage/index.tsx
@@ -1,2 +1,0 @@
-import { AuthorizationPage } from './AuthorizationPage';
-export default AuthorizationPage;
